Install & Configure ESLint & Prettier
Configure ESLint to extend the recommended react configuration,
to use the recommended TypeScript configuration and the recommended
Prettier configuration.

ESLint is also configured to ignore the `src/serviceWorker.js` file
to keep it as close as possible from the native create-react-app
structure.

Prettier is configured to use trailing commas as much as possible
(making the Git history cleaner), single quotes (except in JSX),
lines with up to 80 chars and 2 spaces as tab width.

Resources:

 - [Using ESLint and Prettier in a TypeScript project](https://dev.to/robertcoopercode/using-eslint-and-prettier-in-a-typescript-project-53jb)