Update API query to include deployments
This commit updates the GraphQL API query used to load the
information about the requested repository to also load the latest
deployments.

By default, it will load the latest 5 deployments but this number
can be updated with the `REACT_APP_DEPLOYMENTS` environment
variable.

The query is now loading everything needed to display the info
about a deployment, its commit, author, associated pull request,
status and URL.