Draft of styles of the tab system
This first draft of styles for the Tabs component introduces a
series of small CSS rules for the tab buttons, having a border
with a width depending on the current expansion status and a font
weight depending on the same information.

This commit takes this as an opportunity to test font-weight
transition with variable fonts.