Implement ThemeProvider in BaseStyles
This commit implements the `ThemeProvider` from the
`styled-components` library as part of the `BaseStyles` component
to share the application's theme with all styled components.

To share this theme with all styled components, this commit also
implements a type definition file to override the default one from
`styled-components` to extend the `DefaultTheme` with the typings
from our own `Theme` interface.

References:

 - https://github.com/styled-components/styled-components/issues/1589#issuecomment-456641381
 - https://github.com/castoredc/design-system/blob/bf9fb1d4f918afd0b463163c5406583f857ae8c8/src/typings/styled-components.d.ts