Refactor Tabs component to use React Refs
This commits refactors the Tabs component in order to rely less on
information and Elements coming from the MouseEvent object
available after an click event on a Tab button.

Instead, this component is taking the Tab list provided as prop,
enriches it with a unique React Ref for each panel HTMLElement and
use this unique React Ref as value for the state keeping track of
the active panel.

Using this system, we intend to make the component more readable
and maintainable.