Update TabButton border using pseudo element
Changes the system used to display a thicker bottom border for the
active `TabButton` used by the `Tabs` component.

This component is now using a pseudo element to display and animate
a thicker border when the tab is active without changing the size
of the box, and limiting the impact on all other tabs.