Install Enzyme and create basic test utilities
This commit installs Enzyme as a `devDependency` for this project,
and configures it to be used in Jest snapshots.

This commit also creates a test utility `renderWithApp` rendering
the provided component with Enzyme's render method after wrapping
the provided component in the `TestProvider` one.

Because of that, this commit also contains this new `TestProvider`
component providing all needs for the components to be tested.
At the moment, this `TestProvider` component is activating the
StrictMode and setting up the theme for our tests.