Use react-helmet to manage the title and head info
This commit adds `react-helmet` as a dependency of this project in
order to manage the content of the `<head />` section of the page.

This commit uses this new package to manage and updates the title
of the page (Cerberus during loading then name of the project) and
updates the `meta` element containing the `theme-color` attribute
to use the `tertiary` color of our theme.